Removed old version of kern and replaced name to new one.
