+++ partitura/io/__init__.py
@@ -11,7 +11,7 @@ from .importmidi import load_score_midi, load_performance_midi
 from .musescore import load_via_musescore
 from .importmatch import load_match
 from .importmei import load_mei
-from .importkern_v2 import load_kern
+from .importkern import load_kern
 from .exportkern import save_kern
 from .importparangonada import load_parangonada_csv
 from .exportparangonada import save_parangonada_csv
