+++ tests/test_kern.py
@@ -11,7 +11,7 @@ from tests import KERN_TESTFILES, KERN_TIES, KERN_PATH
 from tempfile import TemporaryDirectory
 from partitura.score import merge_parts
 from partitura.utils import ensure_notearray
-from partitura.io.importkern_v2 import load_kern
+from partitura.io.importkern import load_kern
 from partitura.io.exportkern import save_kern
 from partitura import load_musicxml
 import numpy as np
